[merge] Travis install and other tests fixed. Updating some ingest methods.

+++ test/test_image.py
@@ -155,8 +155,8 @@ class Test_Image_Simple_Catmaid:
     # yz/x/z_y_res
     url = "https://{}/catmaid/{}/{}/yz/{}/{}_{}_{}.png".format(SITE_HOST, p.token, p.channels[0], p.args[0], p.args[4]/512, p.args[2]/512, p.resolution)
     f = getURL (url)
-
-    scale_range = 512*p.voxel[2]/p.voxel[1]
+    
+    scale_range = int(512*p.voxel[2]/p.voxel[1])
     slice_data = np.asarray ( Image.open(StringIO(f.content)) )
     assert ( np.array_equal(slice_data[:scale_range,:], image_data[0,:scale_range,:,0] ))
 
@@ -176,7 +176,7 @@ class Test_Image_Simple_Catmaid:
     url = "https://{}/catmaid/{}/{}/xz/{}/{}_{}_{}.png".format(SITE_HOST, p.token, p.channels[0], p.args[2], p.args[4]/512, p.args[0]/512, p.resolution)
     f = getURL (url)
 
-    scale_range = 512*p.voxel[2]/p.voxel[0]
+    scale_range = int(512*p.voxel[2]/p.voxel[0])
     slice_data = np.asarray ( Image.open(StringIO(f.content)) )
     assert ( np.array_equal(slice_data[:scale_range,:], image_data[0,:scale_range,0,:]) )
